all Methods Implemented

+++ goalapp/src/main/java/com/goalapp/goalapp/exchanges/AddGoalRequest.java
@@ -16,5 +16,5 @@ public class AddGoalRequest {
 
     @NotNull
     @Min(1)
-    private Integer targetAmount;
+    private Long targetAmount;
 }
